fix old refactor issues
